Add `ChildAgenciesDropdown` component to explore data page
